[gh-#1029] refactor and upload other config files
