+++ packages/shared/lib/clients/sync.client.ts
@@ -15,7 +15,7 @@ import {
 } from '../services/activity/activity.service.js';
 import { createSyncJob, updateRunId } from '../services/sync/job.service.js';
 import { getInterval } from '../services/nango-config.service.js';
-import { getSyncConfig } from '../services/sync/config.service.js';
+import { getSyncConfig } from '../services/sync/config/config.service.js';
 import { createSchedule as createSyncSchedule } from '../services/sync/schedule.service.js';
 import connectionService from '../services/connection.service.js';
 import configService from '../services/config.service.js';
