+++ packages/shared/lib/sdk/sync.ts
@@ -1,4 +1,4 @@
-import { getSyncConfigByJobId } from '../services/sync/config.service.js';
+import { getSyncConfigByJobId } from '../services/sync/config/config.service.js';
 import { upsert } from '../services/sync/data/data.service.js';
 import { formatDataRecords } from '../services/sync/data/records.service.js';
 import { createActivityLogMessage } from '../services/activity/activity.service.js';
