+++ packages/shared/lib/services/sync/run.service.unit.test.ts
@@ -1,6 +1,6 @@
 import { expect, describe, it, vi } from 'vitest';
 import SyncRun from './run.service.js';
-import * as ConfigService from './config.service.js';
+import * as ConfigService from './config/config.service.js';
 import environmentService from '../environment.service.js';
 import * as NangoConfigService from '../nango-config.service.js';
 import integationService from './integration.service.js';
